+++ components/card/SimpleCard.jsx
@@ -1,12 +1,13 @@
 import React from 'react'
 import classNames from 'classnames'
 
-const SimpleCard = ({ children, size, style, className, hoverable }) => {
+const SimpleCard = ({ children, size, style, className, hoverable, bordered }) => {
   return (
     <div
       className={classNames('hi-card', 'hi-card--simple', className, {
         [`hi-card--${size}`]: size,
-        'hi-card--hoverable': hoverable
+        'hi-card--hoverable': hoverable,
+        'hi-card--bordered': bordered
       })}
       style={style}
     >