+++ components/form/Item.js
@@ -1,4 +1,4 @@
-import React, { useContext, useState, useEffect, useCallback, useRef } from 'react'
+import React, { useContext, useState, useEffect, useCallback } from 'react'
 import classNames from 'classnames'
 import AsyncValidator from 'async-validator'
 import PropTypes from 'prop-types'
@@ -44,8 +44,6 @@ const FormItem = (props) => {
     listItemValue,
     sort
   } = props
-
-  const FormItemContent = useRef()
 
   const {
     showColon: shouldFormShowColon,
@@ -99,11 +97,6 @@ const FormItem = (props) => {
     [fields]
   )
 
-  let rectCont = FormItemContent.current && FormItemContent.current.getBoundingClientRect()
-
-  useEffect(() => {
-    rectCont = FormItemContent.current && FormItemContent.current.getBoundingClientRect()
-  }, [])
   const resetValidate = useCallback((value = '') => {
     // 清空数据
     setValue(value)
@@ -226,7 +219,7 @@ const FormItem = (props) => {
 
   const labelWidth = useCallback(() => {
     const labelWidth = props.labelWidth || formProps.labelWidth
-    return formProps.labelPosition === 'top'
+    return formProps.labelPlacement === 'top'
       ? '100%'
       : !Number.isNaN(Number(labelWidth))
       ? Number(labelWidth)
@@ -240,7 +233,7 @@ const FormItem = (props) => {
     const _props = componentProps || children.props
     eventName === 'onChange' && _props.onChange && _props.onChange(e, ...args)
     eventName === 'onBlur' && _props.onBlur && _props.onBlur(e, ...args)
-    let value = e.target && e.target.hasOwnProperty(valuePropName) ? e.target[valuePropName] : e
+    let value = e.target && Object.prototype.hasOwnProperty.call(e.target, valuePropName) ? e.target[valuePropName] : e
     if (displayName === 'Counter') {
       value = args[0]
     }
@@ -305,10 +298,9 @@ const FormItem = (props) => {
   obj['hi-form-item--validating'] = validating
   obj['hi-form-item--required'] = isRequired() || required
   const _labelWidth = labelWidth()
-  const contentWidth =
-    formProps.labelPosition !== 'top' && rectCont && rectCont.width ? rectCont.width - _labelWidth : '100%'
+  const contentWidth = formProps.labelPlacement === 'top' ? '100%' : `calc(100% - ${_labelWidth}px)`
   return (
-    <div className={classNames('hi-form-item', className, obj)} style={style} key={field} ref={FormItemContent}>
+    <div className={classNames('hi-form-item', className, obj)} style={style} key={field}>
       {label || label === '' ? (
         <label className="hi-form-item__label" style={{ width: _labelWidth }} key={field + 'label'}>
           {(typeof label === 'string' && label.trim()) || label}
@@ -317,11 +309,7 @@ const FormItem = (props) => {
       ) : (
         <span className="hi-form-item__span" style={{ width: _labelWidth }} key={field + 'label'} />
       )}
-      <div
-        className={'hi-form-item' + '__content'}
-        key={field + '__content'}
-        style={{ width: isNaN(contentWidth) ? '100%' : contentWidth }}
-      >
+      <div className={'hi-form-item' + '__content'} key={field + '__content'} style={{ width: contentWidth }}>
         <div className={'hi-form-item' + '__children'} style={{ alignItems: getItemPosition(contentPosition) }}>
           {renderChildren()}
         </div>