+++ components/select-tree/SelectTreeHook.js
@@ -43,8 +43,13 @@ const SelectTree = ({
   autoExpand,
   overlayClassName,
   theme,
-  localeDatas
+  localeDatas,
+  placeholder: propsPlaceholder,
+  style,
+  optionWidth,
+  placement = 'top-bottom-start'
 }) => {
+  const placeholder = propsPlaceholder || localeDatas.selectTree.placeholder
   const selectedItemsRef = useRef()
   const inputRef = useRef()
   // select 中显示的数量
@@ -97,25 +102,28 @@ const SelectTree = ({
       setNodeEntries(result.nodeEntries)
     }
   }, [data])
-  // 依赖 flattenData & value  解析生成 checkedNodes 或 selectedItems
-  useEffect(() => {
-    if (flattenData.length > 0) {
-      if (type === 'multiple') {
-        const cstatus = parseCheckStatusData(defaultValue.length > 0 ? defaultValue : value, checkedNodes, flattenData)
-        if (cstatus) {
-          setCheckedNodes(cstatus)
-        }
-      } else {
-        const _selectedItems = parseDefaultSelectedItems(defaultValue.length > 0 ? defaultValue : value, flattenData)
-        setSelectedItems(_selectedItems)
-      }
-    }
-  }, [value, flattenData])
   // 当选中项发生改变时(以及首次解析完成默认值后)，更改选中项
   useEffect(() => {
     const _selectedItems = parseSelectedItems(checkedNodes, nodeEntries, showCheckedMode, flattenData)
     setSelectedItems(_selectedItems)
   }, [checkedNodes])
+
+  // 依赖 flattenData & value  解析生成 checkedNodes 或 selectedItems
+  useEffect(() => {
+    if (flattenData.length > 0) {
+      if (type === 'multiple') {
+        const cstatus = parseCheckStatusData(
+          value,
+          value === undefined ? checkedNodes : { checked: value },
+          flattenData
+        )
+        setCheckedNodes(cstatus || { checked: [], semiChecked: [] })
+      } else {
+        const _selectedItems = parseDefaultSelectedItems(value, flattenData)
+        setSelectedItems(_selectedItems)
+      }
+    }
+  }, [value])
 
   // 依赖展开项生成展开节点数据
   useEffect(() => {
@@ -148,6 +156,31 @@ const SelectTree = ({
     }
   }, [selectedItems])
 
+  useEffect(() => {
+    if (data) {
+      const { flattenData = [], nodeEntries } = flattenNodesData(
+        data,
+        defaultExpandIds,
+        defaultExpandAll,
+        type === 'multiple' && showCheckedMode !== 'ALL'
+      )
+      setFlattenData(flattenData)
+      setNodeEntries(nodeEntries)
+      if (flattenData.length > 0) {
+        if (type === 'multiple') {
+          const cstatus = parseCheckStatusData(
+            defaultValue.length > 0 ? defaultValue : value,
+            checkedNodes,
+            flattenData
+          )
+          setCheckedNodes(cstatus || { checked: [], semiChecked: [] })
+        } else {
+          const _selectedItems = parseDefaultSelectedItems(defaultValue.length > 0 ? defaultValue : value, flattenData)
+          setSelectedItems(_selectedItems)
+        }
+      }
+    }
+  }, [])
   // 过滤方法
   const searchTreeNode = (val) => {
     const matchNodes = []
@@ -198,6 +231,7 @@ const SelectTree = ({
       checked: [],
       semiChecked: []
     })
+    onChange && onChange()
   }, [])
 
   /**
@@ -219,12 +253,15 @@ const SelectTree = ({
     return HiRequest({
       ..._dataSource
     }).then((res) => {
-      const nArr = res.data.map((n) => {
-        return {
-          ...n,
-          pId: id
-        }
-      })
+      const { data = [] } = res
+      const nArr =
+        data &&
+        data.map((n) => {
+          return {
+            ...n,
+            pId: id
+          }
+        })
       return nArr
     })
   }, [])
@@ -235,6 +272,7 @@ const SelectTree = ({
    */
   const checkedEvents = (checked, node) => {
     let result = {}
+
     const semiCheckedIds = new Set(checkedNodes.semiChecked)
     const checkedIds = new Set(checkedNodes.checked)
     if (checked) {
@@ -242,10 +280,7 @@ const SelectTree = ({
     } else {
       result = updateUnCheckData(node, flattenData, checkedIds, semiCheckedIds)
     }
-    setCheckedNodes({
-      checked: result.checked,
-      semiChecked: result.semiChecked
-    })
+
     let checkedArr = []
     if (result.checked.length > 0) {
       checkedArr = result.checked.map((id) => {
@@ -257,6 +292,11 @@ const SelectTree = ({
       clearReturnData(checkedArr),
       clearReturnData(node)
     )
+    !value &&
+      setCheckedNodes({
+        checked: result.checked,
+        semiChecked: result.semiChecked
+      })
   }
 
   /**
@@ -326,7 +366,7 @@ const SelectTree = ({
   }
   const searchable = searchMode === 'filter' || searchMode === 'highlight'
   return (
-    <div className={`theme__${theme}`}>
+    <div className={`theme__${theme} hi-selecttree`} style={style}>
       <Trigger
         inputRef={inputRef}
         selectedItemsRef={selectedItemsRef}
@@ -335,6 +375,7 @@ const SelectTree = ({
         selectedItems={selectedItems}
         clearable={clearable}
         show={show}
+        placeholder={placeholder}
         checkedEvents={checkedEvents}
         onTrigger={onTrigger}
         onClear={handleClear}
@@ -343,8 +384,9 @@ const SelectTree = ({
         <Popper
           show={show}
           attachEle={inputRef.current}
-          width={false}
           topGap={5}
+          width={optionWidth}
+          placement={placement}
           overlayClassName={overlayClassName}
           className={`hi-selecttree__popper ${data.length === 0 && dataSource ? 'hi-selecttree__popper--loading' : ''}`}
           onClickOutside={() => setShow(false)}
@@ -431,7 +473,7 @@ SelectTree.defaultProps = {
   onExpand: () => {},
   checkable: false,
   defaultLoadData: true,
-  showCheckedMode: 'PARENT',
+  showCheckedMode: 'ALL',
   defaultExpandAll: false,
   defaultExpandIds: [],
   expandIds: [],