+++ components/select/MultipleInput.js
@@ -27,6 +27,7 @@ const MultipleInput = ({
   const [showCount, setShowCount] = useState(0)
   const tagWrapperRef = useRef('')
   const calShowCountFlag = useRef(true) // 在渲染完成进行测试是否展示 +1
+  const selectedItems = _.uniqBy(cacheSelectItem.concat(propsSelectItem), transKeys(fieldNames, 'id'))
 
   useEffect(() => {
     if (multipleMode === 'nowrap' && calShowCountFlag.current && tagWrapperRef.current) {
@@ -38,7 +39,7 @@ const MultipleInput = ({
       tags.forEach((tag, index) => {
         const tagRect = tag.getBoundingClientRect()
         width += tagRect.width
-        if (width + 90 > tagWrapperRect.width) {
+        if (width + 50 > tagWrapperRect.width && calShowCountFlag.current) {
           // 50是留给显示剩余选项的空间
           calShowCountFlag.current = false
           showCountIndex = index
@@ -54,8 +55,9 @@ const MultipleInput = ({
     e.stopPropagation()
     onClear()
   }
-  const selectedItems = _.uniqBy(cacheSelectItem.concat(propsSelectItem), transKeys(fieldNames, 'id'))
+
   const currentCount = showCount === 0 ? selectedItems.length : showCount
+
   return (
     <div
       className={classNames(