+++ docs/demo/form/section-align.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Form, { LegacyForm } from '../../../components/form'
@@ -34,28 +35,10 @@ const code = [
         return (
           <Form labelWidth='70' labelPlacement='left'>
             <FormItem label='手机号' field='phone'>
-              <Input
-                placeholder={'请输入'}
-                style={{ width: '200' }}
-                onChange={(e, val) => {
-                  setForm({
-                    phone: val
-                  })
-                }}
-              />
+              <Input placeholder={'请输入'} />
             </FormItem>
-            <FormItem label='手机号'>
+            <FormItem label='门店'>
               <Select
-                data={[
-                  { title: '小米2', id: '2-1' },
-                  { title: '小米3', id: '2-2' },
-                  { title: '小米4', id: '2-3' },
-                  { title: '小米5', id: '2-4' },
-                  { title: '电脑', id: '3' },
-                  { title: '笔记本', id: '4' },
-                  { title: '生活周边', id: '5' },
-                  { title: '其它', id: '6' }
-                ]}
                 dataSource={{
                   method: 'GET',
                   key: 'id',