+++ docs/demo/input/section-state.jsx
@@ -4,7 +4,7 @@ import Grid from '../../../components/grid'
 import Input from '../../../components/input'
 import Radio from '../../../components/radio'
 import Button from '../../../components/button'
-const leftOptions = ['基础', '默认值', '禁用', '可清除', '自动聚焦', '手动聚焦']
+const leftOptions = ['基础', '受控', '默认值', '禁用', '可清除', '自动聚焦', '手动聚焦']
 const prefix = 'input-state'
 const desc = '可获取有限长度的字符串，不折行显示'
 const code = [
@@ -24,6 +24,35 @@ class Demo extends React.Component {
   }
 }`,
     opt: ['基础']
+  },
+  {
+    code: `import React from 'react'
+import Grid from '@hi-ui/hiui/es/grid'
+import Radio from '@hi-ui/hiui/es/radio'
+import Input from '@hi-ui/hiui/es/input'\n
+class Demo extends React.Component {
+  constructor (){
+    super()
+    this.state = {
+      value:'HiUI'
+    }
+  }
+  render() {
+    return (
+      <Input
+        style={{ width: 250 }}
+        placeholder='请输入'
+        value={this.state.value}
+        onChange={(e)=>{
+          this.setState({
+            value: e.target.value
+          })
+        }}
+      />
+    )
+  }
+}`,
+    opt: ['受控']
   },
   {
     code: `import React from 'react'
@@ -121,7 +150,7 @@ class Demo extends React.Component {
   render() {
     return (
       <Input
-        style={{ width: 250 }}
+        style={{ width: 250, marginTop:'10px' }}
         placeholder='请输入'
         clearable
       />