+++ docs/demo/modal/section-tip.jsx
@@ -32,40 +32,8 @@ class Demo extends React.Component {
   render(){
     return(
       <div>
-        <Button type="primary" onClick={() => this.setState({visibleError: true})}>打开错误提示</Button>
-        <Button type="primary" onClick={() => this.setState({visibleWarning: true})}>打开警告提示</Button>
-        <Modal
-          title="提示"
-          visible={this.state.visibleError}
-          onCancel={this.cancelEventError.bind(this)}
-          footer={[
-            <Button type="primary" onClick={this.cancelEventError.bind(this)} >我知道了</Button>
-          ]}
-        >
-          <div style={{display: 'flex'}}>
-            <Icon name="close-circle" style={{color: '#EB5252', fontSize: '30px'}} />
-            <div style={{marginLeft: '18px'}}>
-                <span>错误标题</span><br/>
-                <span style={{fontSize: '12px'}}>错误原因错误原因错误原因错误原因错误原因</span>
-            </div>
-          </div>  
-        </Modal>
-        <Modal
-          title="提示"
-          visible={this.state.visibleWarning}
-          onCancel={this.cancelEventWarning.bind(this)}
-          footer={[
-            <Button type="primary" onClick={this.cancelEventWarning.bind(this)} >我知道了</Button>
-          ]}
-        >
-          <div style={{display: 'flex'}}>
-            <Icon name="info-circle" style={{color: '#4284F5', fontSize: '30px'}} />
-            <div style={{marginLeft: '18px'}}>
-              <span>警告标题</span><br/>
-              <span style={{fontSize: '12px'}}>警告具体内容警告具体内容警告具体内容警告具体内容警告具体内容警告具体内容警告具体内容</span>
-            </div>
-          </div>
-        </Modal>
+        <Button type="primary" onClick={() => Modal.confirm({type: 'error', title: '错误', content: <div style={{lineHeight: '48px'}}>操作失败，请联系管理员！</div>, cancelText: null, confirmText:'我知道了'})}>打开错误提示</Button>
+        <Button type="primary" onClick={() => Modal.confirm({type:'warning', title:'警告', content: <div style={{lineHeight: '48px'}}>执行该操作后将无法撤销，是否确定继续？</div>, cancelText: null, confirmText:'确定'})}>打开警告提示</Button>
       </div>
     )
   }